requirements.txt
Flask==3.0.3
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
tensorflow==2.10.0
h5py==3.11.0
Werkzeug==3.0.3